[Bug 1121]: FIXED uint32_t has been replaced by size_t.
git-svn-id: https://svn.oktetlabs.ru/svnroot/prj/oktetlabs/test-environment/src/trunk@30872 9d14d920-75d5-0310-8bd7-a52572981aee

+++ lib/ipc/client.c
@@ -1350,7 +1350,7 @@ ipc_stream_send_message(struct ipc_client *ipcc, const char *server_name,
     if (msg_len + 8 > IPC_TCP_CLIENT_BUFFER_SIZE)
     {
         /* Message is too long to fit into the internal buffer */
-        uint32_t len = msg_len;
+        size_t len = msg_len;
 
         if (write_socket(server->stream.socket,
                          (char *)&len, sizeof(len)) != 0)
@@ -1363,7 +1363,7 @@ ipc_stream_send_message(struct ipc_client *ipcc, const char *server_name,
     }
     else
     {
-        uint32_t len = msg_len;
+        size_t len = msg_len;
 
         memcpy(ipcc->stream.out_buffer, &len, sizeof(len));
         memcpy(ipcc->stream.out_buffer + sizeof(len), msg, msg_len);
